requirements.txt
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
certifi==2026.1.4
cfgv==3.5.0
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.2.1
distlib==0.4.0
Django==4.1.4
django-extensions==3.2.3
djangorestframework==3.14.0
drf-jsonmask==0.3.0
drf-yasg==1.21.4
executing==2.2.1
factory-boy==3.2.1
Faker==40.1.2
filelock==3.20.3
identify==2.6.16
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
ipdb==0.13.9
ipython==9.9.0
ipython_pygments_lexers==1.1.1
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonmask==0.1.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
nodeenv==1.10.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
pluggy==1.6.0
pre-commit==2.20.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pytest==7.2.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
ruamel.yaml==0.19.1
setuptools==80.9.0
sqlparse==0.5.5
stack-data==0.6.3
toml==0.10.2
traitlets==5.14.3
uritemplate==4.2.0
urllib3==2.6.3
virtualenv==20.36.1
wcwidth==0.2.14